+++ Secretaria-war/src/main/java/backingBeans/Expedientes.java
@@ -9,6 +9,7 @@ import javax.ejb.EJB;
 import javax.enterprise.context.RequestScoped;
 import javax.faces.application.FacesMessage;
 import javax.faces.context.FacesContext;
+import javax.inject.Inject;
 import javax.inject.Named;
 
 import excepcionesEJB.ExpedienteException;
@@ -25,6 +26,10 @@ public class Expedientes{
 	
 	@EJB
 	private InterfazExpediente e;
+
+	@Inject
+	private InfoSesion sesion;
+	
 	
 	private Expediente expediente;
 	private PartImpl archivo;
@@ -69,21 +74,31 @@ public class Expedientes{
 	
 	public String buscarTrue() {
 		String respuesta = null;
-		buscar = true;
+		if(sesion.comprobarSesion()) {
+			buscar = true;
+		} else {
+			FacesMessage fm = new FacesMessage("No se ha iniciado sesion");
+            FacesContext.getCurrentInstance().addMessage(null, fm);
+		}
 		return respuesta;
 	}
 	
 	public Expediente buscarExpediente(String referencia) {
 		Expediente expediente = null;
-		try {
-			if(referencia.equals("")) {
-				FacesMessage fm = new FacesMessage("No se ha introducido ninguna referencia");
+		if(sesion.comprobarSesion()) {
+			try {
+				if(referencia.equals("")) {
+					FacesMessage fm = new FacesMessage("No se ha introducido ninguna referencia");
+		            FacesContext.getCurrentInstance().addMessage(null, fm);
+				}else {
+					expediente = e.VisualizarExpediente(Integer.parseInt(referencia));
+				}
+			}catch(ExpedienteException e) {
+				FacesMessage fm = new FacesMessage(e.getMessage());
 	            FacesContext.getCurrentInstance().addMessage(null, fm);
-			}else {
-				expediente = e.VisualizarExpediente(Integer.parseInt(referencia));
 			}
-		}catch(ExpedienteException e) {
-			FacesMessage fm = new FacesMessage(e.getMessage());
+		} else {
+			FacesMessage fm = new FacesMessage("No se ha iniciado sesion");
             FacesContext.getCurrentInstance().addMessage(null, fm);
 		}
 		return expediente;
@@ -91,12 +106,17 @@ public class Expedientes{
 	
 	public List<Expediente> leerDatosAdmin() {
 		List<Expediente> expediente = null;
-		try {
+		if(sesion.comprobarSesion()) {
+			try {
+				
+				expediente = e.mostrarDatosAdmin();
 			
-			expediente = e.mostrarDatosAdmin();
-		
-		}catch(ExpedienteException e) {
-			FacesMessage fm = new FacesMessage("No hay datos que mostrar");
+			}catch(ExpedienteException e) {
+				FacesMessage fm = new FacesMessage("No hay datos que mostrar");
+	            FacesContext.getCurrentInstance().addMessage(null, fm);
+			}
+		} else {
+			FacesMessage fm = new FacesMessage("No se ha iniciado sesion");
             FacesContext.getCurrentInstance().addMessage(null, fm);
 		}
 		return expediente;
@@ -104,10 +124,15 @@ public class Expedientes{
 	
 	public String borrarExpedientes() {
 		String respuesta = null;
-		try {
-			e.borrarExpedientes();
-		} catch(ExpedienteException e) {
-			FacesMessage fm = new FacesMessage(e.getMessage());
+		if(sesion.comprobarSesion()) {
+			try {
+				e.borrarExpedientes();
+			} catch(ExpedienteException e) {
+				FacesMessage fm = new FacesMessage(e.getMessage());
+	            FacesContext.getCurrentInstance().addMessage(null, fm);
+			}
+		} else {
+			FacesMessage fm = new FacesMessage("No se ha iniciado sesion");
             FacesContext.getCurrentInstance().addMessage(null, fm);
 		}
 		return respuesta;
@@ -115,30 +140,35 @@ public class Expedientes{
 	
 	public String importarExpedientes(){
 		String respuesta = null;
-		try {
-			if(archivo.getSubmittedFileName().endsWith(".xlsx")) {
-				String sfile = "/tmp/Expedientes.xlsx";
-				File temporal = new File(sfile);
-				archivo.write(sfile);
-				e.Importar(sfile);
-				temporal.delete();
-				respuesta = "ImportarAdmin.xhtml";
-			} else if(archivo.getSubmittedFileName().endsWith(".csv")) {
-				String sfile = "/tmp/Expedientes.csv";
-				File temporal = new File(sfile);
-				archivo.write(sfile);
-				e.Importar(sfile);
-				temporal.delete();
-				respuesta = "ImportarAdmin.xhtml";
-			} else {
-				FacesMessage fm = new FacesMessage("El archivo no es correcto");
+		if(sesion.comprobarSesion()) {
+			try {
+				if(archivo.getSubmittedFileName().endsWith(".xlsx")) {
+					String sfile = "/tmp/Expedientes.xlsx";
+					File temporal = new File(sfile);
+					archivo.write(sfile);
+					e.Importar(sfile);
+					temporal.delete();
+					respuesta = "ImportarAdmin.xhtml";
+				} else if(archivo.getSubmittedFileName().endsWith(".csv")) {
+					String sfile = "/tmp/Expedientes.csv";
+					File temporal = new File(sfile);
+					archivo.write(sfile);
+					e.Importar(sfile);
+					temporal.delete();
+					respuesta = "ImportarAdmin.xhtml";
+				} else {
+					FacesMessage fm = new FacesMessage("El archivo no es correcto");
+		            FacesContext.getCurrentInstance().addMessage(null, fm);
+				}
+			} catch(ImportarException e){
+				FacesMessage fm = new FacesMessage(e.getMessage());
+	            FacesContext.getCurrentInstance().addMessage(null, fm);
+			} catch (IOException e) {
+				FacesMessage fm = new FacesMessage("Error en el archivo");
 	            FacesContext.getCurrentInstance().addMessage(null, fm);
 			}
-		} catch(ImportarException e){
-			FacesMessage fm = new FacesMessage(e.getMessage());
-            FacesContext.getCurrentInstance().addMessage(null, fm);
-		} catch (IOException e) {
-			FacesMessage fm = new FacesMessage("Error en el archivo");
+		} else {
+			FacesMessage fm = new FacesMessage("No se ha iniciado sesion");
             FacesContext.getCurrentInstance().addMessage(null, fm);
 		}
 		return respuesta;
