+++ Secretaria-war/src/main/java/backingBeans/Matriculas.java
@@ -9,6 +9,7 @@ import javax.ejb.EJB;
 import javax.enterprise.context.RequestScoped;
 import javax.faces.application.FacesMessage;
 import javax.faces.context.FacesContext;
+import javax.inject.Inject;
 import javax.inject.Named;
 
 import excepcionesEJB.ImportarException;
@@ -25,6 +26,10 @@ public class Matriculas{
 	
 	@EJB
 	private InterfazMatricula m;
+
+	@Inject
+	private InfoSesion sesion;
+	
 	
 	private Matricula matricula;
 	private PartImpl archivo;
@@ -87,17 +92,27 @@ public class Matriculas{
 	
 	public String buscarTrue() {
 		String respuesta = null;
-		buscar = true;
+		if(sesion.comprobarSesion()) {
+			buscar = true;
+		} else {
+			FacesMessage fm = new FacesMessage("No se ha iniciado sesion");
+            FacesContext.getCurrentInstance().addMessage(null, fm);
+		}
 		return respuesta;
 	}
 	
 	public Matricula buscarMatricula(String curso, Integer idexp) {
 		Matricula matricula = null;
-		try {
-			LOGGER.info("Buscando Matricula: " + curso + ", " + idexp);
-			matricula = m.VisualizarMatricula(curso,idexp);
-		}catch(MatriculaException e) {
-			FacesMessage fm = new FacesMessage(e.getMessage());
+		if(sesion.comprobarSesion()) {
+			try {
+				LOGGER.info("Buscando Matricula: " + curso + ", " + idexp);
+				matricula = m.VisualizarMatricula(curso,idexp);
+			}catch(MatriculaException e) {
+				FacesMessage fm = new FacesMessage(e.getMessage());
+	            FacesContext.getCurrentInstance().addMessage(null, fm);
+			}
+		} else {
+			FacesMessage fm = new FacesMessage("No se ha iniciado sesion");
             FacesContext.getCurrentInstance().addMessage(null, fm);
 		}
 		return matricula;
@@ -105,12 +120,17 @@ public class Matriculas{
 	
 	public List<Matricula> leerDatosAdmin() {
 		List<Matricula> matricula = null;
-		try {
+		if(sesion.comprobarSesion()) {
+			try {
+				
+				matricula = m.mostrarDatosAdmin();
 			
-			matricula = m.mostrarDatosAdmin();
-		
-		}catch(MatriculaException e) {
-			FacesMessage fm = new FacesMessage("No hay matriculas que mostrar");
+			}catch(MatriculaException e) {
+				FacesMessage fm = new FacesMessage("No hay matriculas que mostrar");
+	            FacesContext.getCurrentInstance().addMessage(null, fm);
+			}
+		} else {
+			FacesMessage fm = new FacesMessage("No se ha iniciado sesion");
             FacesContext.getCurrentInstance().addMessage(null, fm);
 		}
 		return matricula;
@@ -118,10 +138,15 @@ public class Matriculas{
 	
 	public String borrarMatriculas() {
 		String respuesta = null;
-		try {
-			m.borrarMatriculas();
-		} catch(MatriculaException e) {
-			FacesMessage fm = new FacesMessage(e.getMessage());
+		if(sesion.comprobarSesion()) {
+			try {
+				m.borrarMatriculas();
+			} catch(MatriculaException e) {
+				FacesMessage fm = new FacesMessage(e.getMessage());
+	            FacesContext.getCurrentInstance().addMessage(null, fm);
+			}
+		} else {
+			FacesMessage fm = new FacesMessage("No se ha iniciado sesion");
             FacesContext.getCurrentInstance().addMessage(null, fm);
 		}
 		return respuesta;
@@ -129,30 +154,35 @@ public class Matriculas{
 
 	public String importarMatriculas(){
 		String respuesta = null;
-		try {
-			if(archivo.getSubmittedFileName().endsWith(".xlsx")) {
-				String sfile = "/tmp/Matriculas.xlsx";
-				File temporal = new File(sfile);
-				archivo.write(sfile);
-				m.Importar(sfile);
-				temporal.delete();
-				respuesta = "ImportarAdmin.xhtml";
-			} else if(archivo.getSubmittedFileName().endsWith(".csv")) {
-				String sfile = "/tmp/Matriculas.csv";
-				File temporal = new File(sfile);
-				archivo.write(sfile);
-				m.Importar(sfile);
-				temporal.delete();
-				respuesta = "ImportarAdmin.xhtml";
-			} else {
-				FacesMessage fm = new FacesMessage("El archivo no es correcto");
+		if(sesion.comprobarSesion()) {
+			try {
+				if(archivo.getSubmittedFileName().endsWith(".xlsx")) {
+					String sfile = "/tmp/Matriculas.xlsx";
+					File temporal = new File(sfile);
+					archivo.write(sfile);
+					m.Importar(sfile);
+					temporal.delete();
+					respuesta = "ImportarAdmin.xhtml";
+				} else if(archivo.getSubmittedFileName().endsWith(".csv")) {
+					String sfile = "/tmp/Matriculas.csv";
+					File temporal = new File(sfile);
+					archivo.write(sfile);
+					m.Importar(sfile);
+					temporal.delete();
+					respuesta = "ImportarAdmin.xhtml";
+				} else {
+					FacesMessage fm = new FacesMessage("El archivo no es correcto");
+		            FacesContext.getCurrentInstance().addMessage(null, fm);
+				}
+			} catch(ImportarException e){
+				FacesMessage fm = new FacesMessage(e.getMessage());
+	            FacesContext.getCurrentInstance().addMessage(null, fm);
+			} catch (IOException e) {
+				FacesMessage fm = new FacesMessage("Error en el archivo");
 	            FacesContext.getCurrentInstance().addMessage(null, fm);
 			}
-		} catch(ImportarException e){
-			FacesMessage fm = new FacesMessage(e.getMessage());
-            FacesContext.getCurrentInstance().addMessage(null, fm);
-		} catch (IOException e) {
-			FacesMessage fm = new FacesMessage("Error en el archivo");
+		} else {
+			FacesMessage fm = new FacesMessage("No se ha iniciado sesion");
             FacesContext.getCurrentInstance().addMessage(null, fm);
 		}
 		return respuesta;
